New db

+++ backend/src/migrations/createTables.js
@@ -1,13 +1,19 @@
-import { sequelize } from "../models/index.js"; // Import du tunnel de connexion (pour fermeture en fin de script)
-// Le sequelize a "conscience" des différents modèles
-console.log( 'PG_URL in createTables:', process.env.PG_URL );
+import { sequelize } from "../models/index.js"; // Importe l'instance sequelize configurée
+import dotenv from "dotenv"; // Pour charger les variables d’environnement depuis le fichier .env
+dotenv.config(); // Charge les variables d’environnement depuis le fichier .env
 
-console.log("🗑️ Suppression des tables existantes..."); // Notamment pour relancer le script plusieurs fois si on veut faire un reset:db
-await sequelize.drop();
+if (!process.env.PG_URL) {
+  throw new Error("PG_URL is not defined");
+}
 
-console.log("🚧 Définition des tables..."); // Synchroniser le modèle séquelize avec la BDD, ie, RE-CREER la table à partir du modèle Sequelize
-await sequelize.sync();
+console.log( 'PG_URL in createTables:', process.env.PG_URL ); // Pour vérifier que la variable d'environnement est bien chargée
 
-console.log("✅ Migration OK ! Fermeture de la base..."); // On ferme le tunnel de connexion pour que le script s'arrête bien
-await sequelize.close();
+console.log("🗑️ Suppression des tables existantes..."); // Pour relancer le script plusieurs fois si on veut faire un reset:db
+await sequelize.drop(); // Supprime toutes les tables connues par Sequelize (toutes les données disparaissent)
+
+console.log("🚧 Définition des tables..."); // Synchroniser le modèle séquelize avec la BDD, recréer la table à partir du modèle Sequelize
+await sequelize.sync(); // Recrée les tables d’après les définitions des modèles
+
+console.log("✅ Migration OK ! Fermeture de la base..."); // On ferme le tunnel de connexion pour que le script s'arrête correctement
+await sequelize.close(); // Ferme le tunnel de connexion à Postgres (sinon le script risque de "pendre")
